VenturAI/requirements.txt
# Core
python-dotenv
chromadb
requests  # for Grok API calls

# Utils
pandas
numpy

# Dev
pytest
jupyter
